Split domain snapshots into focused modules
