+++ apps/server/tests/domain/test_speed_profile_and_run_suitability.py
@@ -11,7 +11,8 @@ from vibesensor.domain import (
     SpeedProfile,
     SuitabilityCheck,
 )
-from vibesensor.domain.snapshots import DrivingPhaseSummary, SpeedProfileSummary
+from vibesensor.domain.driving_phase_summary import DrivingPhaseSummary
+from vibesensor.domain.speed_profile_summary import SpeedProfileSummary
 from vibesensor.shared.boundaries.diagnostic_case import (
     run_suitability_from_payload,
 )
