+++ apps/server/tests/shared/boundaries/test_golden_file_reconstruction.py
@@ -6,7 +6,8 @@ import pytest
 from test_support import build_speed_sweep_fault_samples, standard_metadata
 
 from vibesensor.domain import SpeedProfile
-from vibesensor.domain.snapshots import DrivingPhaseSummary, SpeedProfileSummary
+from vibesensor.domain.driving_phase_summary import DrivingPhaseSummary
+from vibesensor.domain.speed_profile_summary import SpeedProfileSummary
 from vibesensor.shared.boundaries.diagnostic_case import (
     test_run_from_summary as _reconstruct,
 )
