+++ apps/server/tests/test_support/core.py
@@ -13,7 +13,7 @@ from io import BytesIO
 from typing import Any
 
 from vibesensor.domain import TireSpec
-from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
+from vibesensor.domain.analysis_settings import AnalysisSettingsSnapshot
 from vibesensor.shared.constants import KMH_TO_MPS
 
 _DEFAULT_TIRE = TireSpec.from_aspects(
