+++ apps/server/vibesensor/adapters/simulator/profiles.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 from vibesensor.domain import TireSpec
-from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
+from vibesensor.domain.analysis_settings import AnalysisSettingsSnapshot
 from vibesensor.shared.constants import KMH_TO_MPS
 
 DEFAULT_SPEED_KMH = 100.0
