+++ apps/server/vibesensor/use_cases/run/_recorder_runtime.py
@@ -7,7 +7,7 @@ import logging
 import time
 from typing import TYPE_CHECKING
 
-from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
+from vibesensor.domain.analysis_settings import AnalysisSettingsSnapshot
 from vibesensor.shared.constants import NUMERIC_TYPES
 from vibesensor.shared.ports import ClientTracker, SettingsReader
 from vibesensor.shared.time_utils import utc_now_iso
