+++ docs/ai/repo-map.md
@@ -42,7 +42,7 @@ Scope: detailed file layout, entry points, package structure, and module ownersh
 - `use_cases/run/`: recording pipeline orchestration; `lifecycle_state.py` owns `RunLifecycleState` plus the active-run snapshot, `persistence_writer.py` owns history-write coordination/retry bookkeeping, `sample_flush.py` owns sample-building/flush decisions and auto-stop checks, `status_reporting.py` owns the status/health payload helpers, `_recorder_types.py` owns `RunRecorderConfig`, `RecorderShutdownReport`, and recorder metadata/shutdown helpers, `_recorder_runtime.py` owns the periodic async run loop plus recorder-runtime normalization helpers, `logger.py` owns the thin `RunRecorder` orchestrator, `post_analysis.py` owns the background analysis queue above the injected run-persistence/analysis/error boundary, and `sample_builder.py` owns pure sample-building helpers.
 - `use_cases/updates/`: wheel-based updater workflow orchestration, runtime prerequisite validation, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking. `firmware_cache.py` stays the thin public cache/CLI facade, `firmware_types.py` owns firmware cache/release contracts, `firmware_bundle.py` owns bundle filesystem validation/extraction/metadata helpers, `firmware_release_fetcher.py` owns GitHub firmware HTTP access, `esp_flash_manager.py` owns the flash orchestration state machine, `esp_flash_types.py` owns ESP flash contracts/state/protocols, and `esp_serial.py` / `esp_flash_runner.py` own serial discovery and subprocess execution helpers.
 - `shared/`: cross-cutting typed payloads and ports (`shared/types/` and `shared/ports.py`), boundary serializers/decoders (`shared/boundaries/`), exceptions (`shared/exceptions.py`), JSON helpers (`shared/json_utils.py`), location identifiers (`shared/locations.py`), run-context helpers, and shared pure vehicle-order math (`shared/order_bands.py`). `shared/ports.py` owns the current run-history persistence, active-client lookup, latest-sample/latest-metrics, resolved-speed, and read-only settings ports used by runtime/history/recording flows; `shared/types/` stays focused on payload/model types such as `sensor_frame.py`, `health_snapshot.py`, the feature-scoped Pydantic HTTP contract package `shared/types/api_models/`, and the typed backend/API payloads; `shared/boundaries/run_log.py` owns JSONL run-log decoding/normalization shared by diagnostics and persistence; and `shared/boundaries/diagnostic_case.py` owns persisted-summary decoding plus summary projection, while typed speed and metadata-to-domain factories live on `SpeedProfile`, `Car`, and `Symptom`.
-- `domain/`: DDD-aligned domain model package. Primary domain objects live under `vibesensor/domain/`; larger supporting value objects now split into focused neighboring modules when their behavior grows beyond their parent aggregate file (for example `finding_types.py`, `finding_evidence.py`, `tire_spec.py`, and `order_reference.py`). Domain objects own classification, ranking, actionability, surfacing, lifecycle, and query logic; diagnostics use cases delegate to them. See `docs/domain-model.md` for the full domain object catalog, relationship map, and modeling rules.
+- `domain/`: DDD-aligned domain model package. Primary domain objects live under `vibesensor/domain/`; larger supporting value objects now split into focused neighboring modules when their behavior grows beyond their parent aggregate file (for example `finding_types.py`, `finding_evidence.py`, `tire_spec.py`, `order_reference.py`, `analysis_settings.py`, `run_context.py`, `speed_profile_summary.py`, `driving_phase_summary.py`, and `run_metadata.py`). Domain objects own classification, ranking, actionability, surfacing, lifecycle, and query logic; diagnostics use cases delegate to them. See `docs/domain-model.md` for the full domain object catalog, relationship map, and modeling rules.
 - `apps/ui/src/app/runtime/`: explicit UI runtime owners for shell/chrome state, live transport/payload application, and spectrum/chart orchestration beneath the `UiAppRuntime` composition root.
 - `apps/ui/src/app/features/`: UI feature owners for state updates, API/polling workflows, and delegated event binding across settings, realtime, history, updates, cars, and ESP flash.
 - `apps/ui/src/app/views/`: focused UI renderers and event-target decoders for feature-owned panels; keep string-template DOM assembly here instead of in runtime or feature controller modules.
